Add test_cpucontrol
This commit adds two tests for the cpu control cgroup.

- test_cgcpu_weight

This test spawns two processes, each spawns n_cpu * 16
threads.
The first process is put in cgroup with 8000 cpu shares,
the second in one with 2000 cpu shares.
The threads are started in sync and stopped in sync.
After they have terminated he test checks the parent cgroup
for elapesed time and checks that the cpu usage in that
time is <81% for the first process and <21% for second process.
1% is added for leeway in cpu share availability.

- test_cgcpu_weight_threaded

This test spawns n_cpu * 32 threads. Half of them are put
in a cgroup with 8000 cpu shares, half in one with 2000 cpu
shares. The threads are started and stopped in sync.
When the group1 have terminated, the test checks
the parent cgroup for elapesed time and checks that the cpu
usage in that time for the first group is <81% and <21% for
the second group.
3% is added for leeway in cpu share availability